feat: rename & relocation SignToDocument component
